requirements.txt
firecrawl>=4.0.0
PyYAML>=6.0.1
pytest>=8.0.0
requests>=2.31.0
